requirements.txt
pytorch == 1.12.0
numpy == 1.22.3
scipy == 1.7.3
scikit-learn == 1.1.2
Pillow == 9.2.0
wandb == 0.13.1
termcolor == 1.1.0
opencv-python == 4.6.0
torchsummary == 1.5.1